# api-gateway/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Last Mile Delivery API
  version: 1.0.0
  description: >
    API documentation for Last Mile Delivery Backend.
    The system uses a microservice-lite architecture with API Gateway.
servers:
  - url: http://localhost:3000
    description: Local API Gateway
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderItemRequest:
      type: object
      required:
        - itemName
      properties:
        itemName:
          type: string
          example: Shoes
        quantity:
          type: integer
          example: 1
        weightKg:
          type: number
          nullable: true
          example: 0.8
        note:
          type: string
          nullable: true
          example: Fragile item
...
    CreateOrderRequest:
      type: object
      required:
        - pickupAddressLine
        - receiverName
        - receiverPhone
        - deliveryAddressLine
        - items
      properties:
        pickupAddressLine:
          type: string
          example: 12 Nguyen Van Cu
        pickupWard:
          type: string
          example: Ward 1
        pickupDistrict:
          type: string
          example: District 5
        pickupCity:
          type: string
          example: Ho Chi Minh City
        pickupLat:
          type: number
          nullable: true
          example: 10.762622
        pickupLng:
          type: number
          nullable: true
          example: 106.660172
        receiverName:
          type: string
          example: Tran Van B
        receiverPhone:
          type: string
          example: "0912345678"
        deliveryAddressLine:
          type: string
          example: 25 Le Loi
        deliveryWard:
          type: string
          example: Ben Nghe
        deliveryDistrict:
          type: string
          example: District 1
        deliveryCity:
          type: string
          example: Ho Chi Minh City
        deliveryLat:
          type: number
          nullable: true
          example: 10.776889
        deliveryLng:
          type: number
          nullable: true
          example: 106.700806
        distanceKm:
          type: number
          example: 5.2
        shippingFee:
          type: number
          nullable: true
          example: 41000
        paymentMethod:
          type: string
          enum:
            - COD
            - BANK_TRANSFER
            - MOMO
            - VNPAY
          example: COD
        note:
          type: string
          nullable: true
          example: Please call before delivery
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemRequest"
...
    CreatePaymentRequest:
      type: object
      required:
        - orderId
        - amount
        - paymentMethod
      properties:
        orderId:
          type: string
          format: uuid
        amount:
          type: number
          example: 41000
        paymentMethod:
          type: string
          enum:
            - COD
            - BANK_TRANSFER
            - MOMO
            - VNPAY
          example: COD
        customerId:
          type: string
          format: uuid
          nullable: true
        providerTransactionId:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
...
    CreateNotificationRequest:
      type: object
      required:
        - userId
        - title
        - message
      properties:
        userId:
          type: string
          format: uuid
        title:
          type: string
          example: Order updated
        message:
          type: string
          example: Your order status has changed.
        type:
          type: string
          enum:
            - ORDER
            - PAYMENT
            - DRIVER
            - DISPATCH
            - SYSTEM
            - PROMOTION
          example: ORDER
        channel:
          type: string
          enum:
            - IN_APP
            - EMAIL
            - SMS
            - PUSH
          example: IN_APP
        metadata:
          type: object
          nullable: true
...
    AdminCreateUserRequest:
      type: object
      required:
        - fullName
        - phone
        - password
        - role
      properties:
        fullName:
          type: string
          example: Admin Created User
        phone:
          type: string
          example: "0900000099"
        email:
          type: string
          nullable: true
          example: created@test.com
        password:
          type: string
          example: "123456"
        role:
          type: string
          enum:
            - ADMIN
            - CUSTOMER
            - DRIVER
          example: CUSTOMER
...
paths:
...
  /orders:
    post:
      tags:
        - Orders
      summary: Create a new order
      description: CUSTOMER or ADMIN can create an order.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created successfully
        "400":
          description: Invalid order data
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

    get:
      tags:
        - Orders
      summary: Get all orders
      description: ADMIN only.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          schema:
            type: integer
            example: 10
        - name: status
          in: query
          schema:
            type: string
            example: PENDING
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order list
        "401":
          description: Unauthorized
        "403":
          description: Admin role required
...
  /payments:
    post:
      tags:
        - Payments
      summary: Create payment transaction
      description: CUSTOMER can create payment for own order. ADMIN can create payment for valid order.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Payment created successfully
        "400":
          description: Invalid payment data
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Order not found
...
  /drivers:
    get:
      tags:
        - Drivers
      summary: Get driver list
      description: ADMIN only.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          schema:
            type: integer
            example: 10
        - name: status
          in: query
          schema:
            type: string
            example: ONLINE
        - name: verificationStatus
          in: query
          schema:
            type: string
            example: APPROVED
      responses:
        "200":
          description: Driver list
        "401":
          description: Unauthorized
        "403":
          description: Admin role required
...
  /notifications:
    post:
      tags:
        - Notifications
      summary: Create notification
      description: ADMIN only.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationRequest"
      responses:
        "201":
          description: Notification created
        "400":
          description: Invalid notification data
        "401":
          description: Unauthorized
        "403":
          description: Admin role required

    get:
      tags:
        - Notifications
      summary: Get my notifications
      description: CUSTOMER, DRIVER, or ADMIN.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          schema:
            type: integer
            example: 10
        - name: status
          in: query
          schema:
            type: string
            enum:
              - UNREAD
              - READ
        - name: type
          in: query
          schema:
            type: string
            enum:
              - ORDER
              - PAYMENT
              - DRIVER
              - DISPATCH
              - SYSTEM
              - PROMOTION
      responses:
        "200":
          description: My notifications
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: Get user list
      description: ADMIN only.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          schema:
            type: integer
            example: 10
        - name: role
          in: query
          schema:
            type: string
            enum:
              - ADMIN
              - CUSTOMER
              - DRIVER
        - name: status
          in: query
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - DELETED
        - name: search
          in: query
          schema:
            type: string
            example: "0900000001"
      responses:
        "200":
          description: User list
        "401":
          description: Unauthorized
        "403":
          description: Admin role required

    post:
      tags:
        - Admin
      summary: Create user by admin
      description: ADMIN only.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateUserRequest"
      responses:
        "201":
          description: User created
        "400":
          description: Invalid user data
        "401":
          description: Unauthorized
        "403":
          description: Admin role required
        "409":
          description: Phone or email already exists
...
  /admin/orders:
    get:
      tags:
        - Admin
      summary: Get all orders for admin
      description: ADMIN only.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          schema:
            type: integer
            example: 10
      responses:
        "200":
          description: Admin order list
        "401":
          description: Unauthorized
        "403":
          description: Admin role required
...
  /admin/drivers:
    get:
      tags:
        - Admin
      summary: Get all driver profiles for admin
      description: ADMIN only.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          schema:
            type: integer
            example: 10
      responses:
        "200":
          description: Admin driver list
        "401":
          description: Unauthorized
        "403":
          description: Admin role required
...
  /admin/payments:
    get:
      tags:
        - Admin
      summary: Get all payments for admin
      description: ADMIN only.
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          schema:
            type: integer
            example: 10
      responses:
        "200":
          description: Admin payment list
        "401":
          description: Unauthorized
        "403":
          description: Admin role required
...